Add an ignore_mask argument to ignore certain source pixels
In practice, the resampling process is not always entirely linear: After
rescaling, pixel values might be clamped to their value range. For
example, when a white image is resampled with zero-padding and the
kernel has negative lobes, the second and third pixel lines from the
edge will be clamped.

When descaling, these clamped pixels give wrong information, so they
will negatively affect the descale result. In the above example, they
will cause dirty borders in the descaled image.

One idea of fixing this problem is to simply discard any pixels which
are suspected to be clipped from the linear system of equations. Since
the system of equations is overdetermined, this can still give an
accurate descale result, as long as no longer contiguous stretches of
pixels are discarded.

This commit adds an ignore_mask argument to the descale functions.
Pixels selected by this mask will have their corresponding equations
dropped from the system of equations. This requires partially
recomputing the matrix and its LDLT decomposition whenever this mask
changes, but this is still reasonably efficient since the matrix is
banded.

The ignore_mask is only supported when only descaling along one axis:
When descaling along both axes, a second mask at an intermediate
resolution would be required for the second descale, so this is better
off being left to the user.